backend/pyproject.toml
[tool.poetry]
name = "need-cv-asap"
version = "0.1.0"
description = "AI CV Analysis Tool"
authors = ["Your Name <you@example.com>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
python-multipart = "^0.0.6"
pydantic = "^2.4.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pymongo = "^4.5.0"
python-dotenv = "^1.0.0"
pdf2image = "^1.16.3"
pytesseract = "^0.3.10"
spacy = "^3.7.2"
transformers = "^4.35.0"
torch = "^2.1.0"
openai = "^1.61.1"
pypdf2 = "^3.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.10.1"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
